Don't try to use fsnotify on mobile or playground
